requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
numpy
faiss-cpu
sentence-transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
